+++ CHANGELOG.md
@@ -13,6 +13,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Updated to use wdfn-viz which updates the USWDS version to 2.2.1
 - Only selected description's time series will now appear in the hydrograph's tooltip.
 - Include the selected method's description, if defined, in the hydrograph title.
+- Monitoring location points will now render on top of the upstream/downstream lines on the map
 
 ## [0.20.0](https://github.com/usgs/waterdataui/compare/waterdataui-0.19.0...waterdataui-0.20.0) - 2019-10-04
 ### Added